Move FusionKernelRuntime to its own file.
