+++ csrc/python_frontend/fusion_cache.cpp
@@ -10,6 +10,7 @@
 
 #include <debug.h>
 #include <instrumentation.h>
+#include <kernel_cache_runtime.h>
 #include <options.h>
 #include <python_frontend/fusion_cache.h>
 #include <serde/fusion_record.h>
